redefine binomial coefficients independently of polynomials
In this comment, (m choose n) is redefined directly as a fraction,
rather than as an evaluation of the polynomial (X choose n).
This should make most of the chapter (except for some alternative
proofs and a few exercises) independent of the definition of
polynomials (and thus more accessible).

Several theorems and propositions had to be reshuffled for this
to work, and the two proofs of the Vandermonde convolution
identity have traded places.
Also, a few typos have been removed and a few details added.
And a few footnotes have been moved into the body text.