add chapter on mathematical induction (\ref{chp.ind})
This chapter is meant to be an introduction to induction,
but also to provide proofs for several fundamental results
(e.g., the well-definedness of a sum of n numbers -- i.e.,
it doesn't depend on the order) that are used later on.
At this point it has no exercises, but I may add some in
the future.

Also, there is now a proper introduction.